+++ src/use-lunatic/commons/fill-components/fill-component-expressions.ts
@@ -47,7 +47,8 @@ type UntranslatedProperties =
 	| 'controls'
 	| 'conditionFilter'
 	| 'conditionReadOnly'
-	| 'components';
+	| 'components'
+	| 'optionFilter';
 export type DeepTranslateExpression<T> = T extends LunaticExpression
 	? ReactNode
 	: T extends (infer ElementType)[]